+++ grafana-plugin/src/assets/style/vars.css
@@ -1,18 +1,8 @@
 :root {
-  --maintenance-background: repeating-linear-gradient(45deg, #f6ba52, #f6ba52 20px, #ffd180 20px, #ffd180 40px);
-  --green-5: #6ccf8e;
   --green-6: #73d13d;
-  --red-5: #ff4d4f;
-  --orange-5: #ffa940;
-  --warning-1: #f8d06b;
-  --blue-2: #bae7ff;
-  --gray-5: #d9d9d9;
-  --gray-8: #595959;
   --gray-9: #434343;
   --cyan-1: #e6fffb;
-  --purple-9: #22075e;
   --border-radius: 2px;
-  --gradient-brandHorizontal: linear-gradient(90deg, #f83 0%, #f53e4c 100%);
   --gradient-brandVertical: linear-gradient(0.01deg, #f53e4c -31.2%, #f83 113.07%);
   --always-gray: #ccccdc;
   --title-marginBottom: 16px;
@@ -28,7 +18,6 @@
 .theme-light {
   --cards-background: rgba(244, 245, 245);
   --highlighted-row-bg: var(--cyan-1);
-  --disabled-button-color: #bdbdbd;
   --primary-background: rgb(255, 255, 255);
   --secondary-background: rgb(244, 245, 245);
   --border: 1px solid rgba(36, 41, 46, 0.12);
@@ -40,7 +29,6 @@
   --error-text-color: #ff5286;
   --primary-text-link: #1f62e0;
   --timeline-icon-background: rgba(70, 76, 84, 0);
-  --timeline-icon-background-resolution-note: rgba(50, 116, 217, 0);
   --oncall-icon-stroke-color: #fff;
   --hover-selected: #f4f5f5;
   --background-canvas: #f4f5f5;
@@ -50,8 +38,6 @@
   --border-medium: 1px solid rgba(36, 41, 46, 0.3);
   --border-strong: 1px solid rgba(36, 41, 46, 0.4);
   --border-weak: 1px solid rgba(36, 41, 46, 0.12);
-  --shadows-z1: 0 1px 2px rgba(24, 26, 27, 0.2);
-  --shadows-z2: 0 4px 8px rgba(24, 26, 27, 0.2);
   --shadows-z3: 0 13px 20px 1px rgba(24, 26, 27, 0.18);
   --tag-background-primary: rgba(50, 116, 217, 0.15);
   --tag-border-primary: rgb(136, 174, 233);
@@ -99,8 +85,6 @@
   --border-medium: 1px solid rgba(204, 204, 220, 0.15);
   --border-strong: 1px solid rgba(204, 204, 220, 0.25);
   --border-weak: 1px solid rgba(204, 204, 220, 0.07);
-  --shadows-z1: 0 1px 2px rgba(24, 26, 27, 0.75);
-  --shadows-z2: 0 4px 8px rgba(24, 26, 27, 0.75);
   --shadows-z3: 0 8px 24px rgb(1, 4, 9);
   --tag-background-primary: rgba(87, 148, 242, 0.15);
   --tag-border-primary: rgb(13, 72, 163);
@@ -114,8 +98,6 @@
   --tag-background-danger: rgba(242, 73, 92, 0.15);
   --tag-border-danger: rgb(151, 11, 27);
   --tag-text-danger: rgb(247, 144, 156);
-  --button-background: rgba(204, 204, 220, 0.1);
-  --button-hover-background: rgba(204, 204, 220, 0.14);
   --box-background: rgba(10, 10, 10, 0.4);
   --working-hours-shades-color: rgba(17, 18, 23, 0.15);
   --working-hours-shades-color-light: rgba(17, 18, 23, 0.1);
