+++ grafana-plugin/src/components/Avatar/Avatar.tsx
@@ -3,7 +3,7 @@ import React, { FC } from 'react';
 import { cx } from '@emotion/css';
 import { useStyles2 } from '@grafana/ui';
 
-import { bem } from 'utils/utils';
+import { bem } from 'styles/utils.styles';
 
 import { getAvatarStyles } from './Avatar.styles';
