+++ grafana-plugin/src/components/GBlock/Block.tsx
@@ -3,7 +3,7 @@ import React, { FC, HTMLAttributes } from 'react';
 import { cx } from '@emotion/css';
 import { useStyles2 } from '@grafana/ui';
 
-import { bem } from 'utils/utils';
+import { bem } from 'styles/utils.styles';
 
 import { getBlockStyles } from './Block.styles';
