+++ grafana-plugin/src/components/Policy/DragHandle.tsx
@@ -4,7 +4,7 @@ import { cx } from '@emotion/css';
 import { Icon, useStyles2 } from '@grafana/ui';
 import { SortableHandle } from 'react-sortable-hoc';
 
-import { bem } from 'utils/utils';
+import { bem } from 'styles/utils.styles';
 
 import { getPolicyStyles } from './Policy.styles';
