+++ grafana-plugin/src/components/Policy/PolicyNote.tsx
@@ -4,7 +4,7 @@ import { cx } from '@emotion/css';
 import { Icon, Tooltip, useStyles2 } from '@grafana/ui';
 import MediaQuery from 'react-responsive';
 
-import { bem } from 'utils/utils';
+import { bem } from 'styles/utils.styles';
 
 import { getPolicyStyles } from './Policy.styles';
