--- grafana-plugin/src/components/Tag/Tag.module.css
@@ -1,12 +1,0 @@
-.root {
-  border-radius: 2px;
-  line-height: 100%;
-  padding: 5px 8px;
-  display: inline-block;
-  white-space: nowrap;
-}
-
-.size-small {
-  font-size: 12px;
-  height: 24px;
-}
