+++ grafana-plugin/src/components/Tag/Tag.tsx
@@ -1,13 +1,11 @@
 import React, { FC } from 'react';
 
-import { css } from '@emotion/css';
+import { css, cx } from '@emotion/css';
 import { GrafanaTheme2 } from '@grafana/data';
 import { useStyles2 } from '@grafana/ui';
 import cn from 'classnames/bind';
 
-import { bem } from 'utils/utils';
-
-import styles from 'components/Tag/Tag.module.css';
+import { bem } from 'styles/utils.styles';
 
 interface TagProps {
   color?: string;
@@ -19,8 +17,6 @@ interface TagProps {
   forwardedRef?: React.MutableRefObject<HTMLSpanElement>;
   size?: 'small' | 'medium';
 }
-
-const cx = cn.bind(styles);
 
 export const Tag: FC<TagProps> = (props) => {
   const styles = useStyles2(getStyles);
@@ -34,7 +30,7 @@ export const Tag: FC<TagProps> = (props) => {
   return (
     <span
       style={style}
-      className={cx(styles.root, bem(styles.root, size), className, 'test-class-here')}
+      className={cx(styles.root, bem(styles.root, size), className)}
       onClick={onClick}
       ref={props.forwardedRef}
     >
@@ -49,6 +45,7 @@ const getStyles = (_theme: GrafanaTheme2) => {
       border-radius: 2px;
       line-height: 100%;
       padding: 5px 8px;
+      color: #fff;
       display: inline-block;
       white-space: nowrap;
     `,
