+++ grafana-plugin/src/components/TooltipBadge/TooltipBadge.tsx
@@ -4,7 +4,7 @@ import { cx } from '@emotion/css';
 import { Icon, Tooltip, IconName, VerticalGroup, HorizontalGroup, useStyles2 } from '@grafana/ui';
 
 import { Text, TextType } from 'components/Text/Text';
-import { bem } from 'utils/utils';
+import { bem } from 'styles/utils.styles';
 
 import { getTooltipBadgeStyles } from './TooltipBadge.styles';
